requirements.txt
fastapi==0.95.0
uvicorn==0.18.0
openai==1.93.0
python-dotenv==1.0.1
requests==2.28.1
pydantic==1.10.2
streamlit==1.16.0
langchain==0.0.149
